Task6 # JWT authorization and CORS

+++ api/controllers/userController.js
@@ -1,10 +1,11 @@
 import { userService } from '../services/userService';
+const createError = require('http-errors');
 
 const create = async (req, res, next) => {
     try {
         const user = await userService.getUserByLogin(req.body.login);
         if (user) {
-            res.status(409).send(`User login ${req.body.login} already exists!`);
+            throw createError.Conflict(`User login ${req.body.login} already exists!`);
         } else {
             const userData = await userService.createUser(req.body);
             res.status(201).json(userData.id);
@@ -29,7 +30,7 @@ const getById = async (req, res, next) => {
         if (user) {
             res.json(user);
         } else {
-            res.status(404).send(`User with id ${req.params.id} not found!`);
+            throw createError.NotFound(`User id ${req.params.id} not found!`);
         }
     } catch (error) {
         return next(error);
@@ -39,12 +40,12 @@ const getById = async (req, res, next) => {
 const update = async (req, res, next) => {
     try {
         const user = await userService.getUserById(req.params.id);
-        if (!user) {
-            res.status(404).send(`User with id ${req.params.id} not found!`);
-        } else {
+        if (user) {
             await userService.updateUser(req.params.id, req.body);
             await user.reload();
             res.json(user);
+        } else {
+            throw createError.NotFound(`User id ${req.params.id} not found!`);
         }
     } catch (error) {
         return next(error);
@@ -53,8 +54,13 @@ const update = async (req, res, next) => {
 
 const remove = async (req, res, next) => {
     try {
-        await userService.deleteUserById(req.params.id);
-        res.status(200).end();
+        const user = await userService.getUserById(req.params.id);
+        if (user) {
+            await userService.deleteUserById(req.params.id);
+            res.status(200).end();
+        } else {
+            throw createError.NotFound(`User id ${req.params.id} not found!`);
+        }
     } catch (error) {
         return next(error);
     }
